fix: login logic
