+++ src/hooks/useLogin.js
@@ -1,24 +1,34 @@
+import { useEffect } from "react";
 import { useCookies } from 'react-cookie';
 import { loginApi } from '../api/authApi';
+import { useNavigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
 function useLogin() {
-  const [, setCookie] = useCookies(['accessToken', 'refreshToken', 'isActive']);
+  const [cookies, setCookie] = useCookies(['accessToken', 'refreshToken', 'isActive']);
+  const navigate = useNavigate();
+  const code = new URLSearchParams(window.location.search).get('code');
+  const { option } = useParams(); 
 
-  const login = async (code, option) => {
+  useEffect(() => {
     if (code && option) {
-      loginApi(code, option)
-        .then((response) => {
-          setCookie("accessToken", response.data.accessToken, { maxAge: 2 * 60 * 60 });
-          setCookie("refreshToken", response.data.refreshToken, { maxAge: 24 * 7 * 60 });
-          setCookie('isActive', response.data.registerStateEnum, { maxAge: 2 * 60 * 60 });
-        })
-        .catch((error) => {
-          console.log(error.message);
-        });
+        loginApi(code, option)
+            .then((response) => {
+            setCookie("accessToken", response.data.accessToken, { path: '/', maxAge: 2 * 60 * 60 });
+            setCookie("refreshToken", response.data.refreshToken, { path: '/', maxAge: 24 * 7 * 60 });
+            setCookie('isActive', response.data.registerStateEnum, { path: '/', maxAge: 2 * 60 * 60 });
+            })
+            .catch((error) => {
+            console.log(error.message);
+            });
     }
-  };
+  }, [code, option]);
 
-  return { login };
+  useEffect(() => {
+    if (cookies.accessToken) {
+      navigate('/chatbot');
+    }
+  }, [cookies]);
 }
 
 export default useLogin;
